\name{logLik}
\alias{logLik.metaplus}
\title{log Likelikelihood for metaplus object}
\description{
Returns the log Likelihood for a metaplus object.
}
\usage{
\method{logLik}{metaplus}(object, \ldots)
}
\arguments{
  \item{object}{metaplus object}
  \item{\dots}{ additional argument; currently none are used. }
}
\value{
The loglikelihood of teh fiutted model.
}
\author{Ken Beath}
\keyword{methods}